golangci-lint round #3
this is the third round of preparing to use the golangci-lint on our
code base.

+++ libpod/stats.go
@@ -86,7 +86,7 @@ func getMemLimit(cgroupLimit uint64) uint64 {
 		return cgroupLimit
 	}
 
-	physicalLimit := uint64(si.Totalram)
+	physicalLimit := si.Totalram
 	if cgroupLimit > physicalLimit {
 		return physicalLimit
 	}
